requirements-dev.txt
-r requirements.txt
beautifulsoup4==4.5.1
django-webtest==1.8.0
Fabric==1.12.0
factory-boy==2.7.0
path.py==8.2.1
pyquery==1.2.15
six==1.10.0
waitress==1.0.0
WebOb==1.6.1
WebTest==2.0.23
